cosmwasm: fix check of activated height in CosmWasm controller (#489)
This PR fixes a bug in CosmWasm consumer controller where we need to
check whether the unmarshalled activated height is 0 or not. If it's 0
then it means we are unmarshalling something fishy and actually BTC
staking is not activated, and we need to return an error, as
[expected](https://github.com/babylonchain/finality-provider/blob/fee9b10e886a098ebd97b7ccfcf9e31bf465a746/finality-provider/service/chain_poller.go#L211-L220)
by RPC poller. This fix allows us to remove some hacks and makes e2e /
local deployment pass.